Add handlebars helpers
