+++ test/handlebars-helpers.test.ts
@@ -1,0 +1,8 @@
+import { ifHasString } from '../src/handlebars-helpers';
+
+describe('handlebars-helper', () => {
+  it('should provide', () => {
+    const actual = ifHasString('stuff', 'blue', '', options);
+    expect(actual).toBeDefined()
+  });
+});
